Cast preprocessor blob to ONNX input dtype
The preprocessor always produces a float32 blob, but FP16-exported ONNX
models declare a float16 input and ORT rejects the dtype mismatch with
"Unexpected input data type. Actual: (tensor(float)), expected:
(tensor(float16))". Read the session's declared input dtype once at init
and cast in `_run_inference` only when it differs (no-op for FP32).

Also extend the existing FP16 export test to actually run inference
through the exported model — previously it only checked the declared
input dtype on disk, which is how this regression slipped through.

+++ libreyolo/models/yolo9/nn.py
@@ -536,8 +536,10 @@ class DDetect(nn.Module):
 
         for i, stride in enumerate(strides):
             _, _, h, w = feats[i].shape
-            sx = torch.arange(end=w, device=device, dtype=dtype) + grid_cell_offset
-            sy = torch.arange(end=h, device=device, dtype=dtype) + grid_cell_offset
+            # ONNX's Range op rejects fp16 inputs, so do the integer counting
+            # in fp32 and then cast to the model dtype. No effect on fp32 runs.
+            sx = torch.arange(end=w, device=device, dtype=torch.float32).to(dtype) + grid_cell_offset
+            sy = torch.arange(end=h, device=device, dtype=torch.float32).to(dtype) + grid_cell_offset
             sy, sx = torch.meshgrid(sy, sx, indexing="ij")
             anchor_points.append(torch.stack((sx, sy), -1).view(-1, 2))
             stride_tensor.append(
